+++ gitman/cli.py
@@ -7,13 +7,12 @@ import argparse
 import logging
 
 from . import CLI, VERSION, DESCRIPTION
-from . import common
-from . import commands
+from . import common, exceptions, commands
 
 log = logging.getLogger(__name__)
 
 
-def main(args=None, function=None):
+def main(args=None, function=None):  # pylint: disable=too-many-statements
     """Process command-line arguments and run the program."""
 
     # Shared options
@@ -43,6 +42,11 @@ def main(args=None, function=None):
                                      parents=[debug], **shared)
     subs = parser.add_subparsers(help="", dest='command', metavar="<command>")
 
+    # Init parser
+    info = "create a new configuration file for the project"
+    sub = subs.add_parser('init', description=info.capitalize() + '.',
+                          help=info, parents=[debug], **shared)
+
     # Install parser
     info = "get the specified versions of all dependencies"
     sub = subs.add_parser('install', description=info.capitalize() + '.',
@@ -70,7 +74,7 @@ def main(args=None, function=None):
                        action='store_false', dest='lock', default=None,
                        help="disable recording of versions for later reinstall")
 
-    # Display parser
+    # List parser
     info = "display the current version of each dependency"
     sub = subs.add_parser('list', description=info.capitalize() + '.',
                           help=info, parents=[debug, project, depth], **shared)
@@ -115,19 +119,22 @@ def main(args=None, function=None):
     common.configure_logging(namespace.verbose)
 
     # Run the program
-    function, args, kwargs, exit_message = _get_command(function, namespace)
-    if function is None:
+    function, args, kwargs = _get_command(function, namespace)
+    if function:
+        _run_command(function, args, kwargs)
+    else:
         parser.print_help()
         sys.exit(1)
-    _run_command(function, args, kwargs, exit_message)
 
 
-def _get_command(function, namespace):
+def _get_command(function, namespace):  # pylint: disable=too-many-statements
     args = []
     kwargs = {}
-    exit_message = None
 
-    if namespace.command in ['install', 'update']:
+    if namespace.command == 'init':
+        function = commands.init
+
+    elif namespace.command in ['install', 'update']:
         function = getattr(commands, namespace.command)
         args = namespace.name
         kwargs.update(root=namespace.root,
@@ -139,7 +146,6 @@ def _get_command(function, namespace):
         if namespace.command == 'update':
             kwargs.update(recurse=namespace.recurse,
                           lock=namespace.lock)
-        exit_message = "Run again with '--force' to overwrite changes"
 
     elif namespace.command == 'list':
         function = commands.display
@@ -156,7 +162,6 @@ def _get_command(function, namespace):
         function = commands.delete
         kwargs.update(root=namespace.root,
                       force=namespace.force)
-        exit_message = "Run again with '--force' to ignore changes"
 
     elif namespace.command == 'show':
         function = commands.show
@@ -171,29 +176,41 @@ def _get_command(function, namespace):
         function = commands.edit
         kwargs.update(root=namespace.root)
 
-    return function, args, kwargs, exit_message
+    return function, args, kwargs
 
 
-def _run_command(function, args, kwargs, exit_message):
+def _run_command(function, args, kwargs):
     success = False
+    exit_message = None
     try:
         log.debug("Running %s command...", getattr(function, '__name__', 'a'))
         success = function(*args, **kwargs)
     except KeyboardInterrupt:
         log.debug("Command canceled")
-    except RuntimeError as exc:
-        common.dedent(0)
-        common.show(str(exc), color='error')
-        common.show()
+    except exceptions.UncommittedChanges as exception:
+        _show_error(exception)
+        exit_message = "Run again with '--force' to discard changes"
+    except exceptions.ScriptFailure as exception:
+        _show_error(exception)
+        exit_message = "Run again with '--force' to ignore script errors"
+    finally:
         if exit_message:
             common.show(exit_message, color='message')
-            common.show()
+            common.newline()
 
     if success:
         log.debug("Command succeeded")
     else:
         log.debug("Command failed")
         sys.exit(1)
+
+
+def _show_error(exception):
+    # TODO: require level=, evaluate all calls to dedent()
+    common.dedent(0)
+    common.newline()
+    common.show(str(exception), color='error')
+    common.newline()
 
 
 if __name__ == '__main__':  # pragma: no cover (manual test)