+++ gitman/tests/test_cli.py
@@ -8,6 +8,7 @@ from expecter import expect
 
 from gitman import cli
 from gitman.common import _Config
+from gitman.exceptions import UncommittedChanges, ScriptFailure
 
 
 class TestMain:
@@ -44,7 +45,20 @@ class TestMain:
     def test_main_error(self):
         """Verify runtime errors are handled."""
         with pytest.raises(SystemExit):
-            cli.main([], Mock(side_effect=RuntimeError))
+            cli.main([], Mock(side_effect=UncommittedChanges))
+        with pytest.raises(SystemExit):
+            cli.main([], Mock(side_effect=ScriptFailure))
+
+
+class TestInit:
+    """Unit tests for the `init` command."""
+
+    @patch('gitman.commands.init')
+    def test_install(self, mock_init):
+        """Verify the 'install' command can be run."""
+        cli.main(['init'])
+
+        mock_init.assert_called_once_with()
 
 
 class TestInstall: