+++ gitman/tests/test_models_source.py
@@ -17,31 +17,33 @@ class TestSource:
 
     def test_init_defaults(self):
         """Verify a source has a default revision."""
-        source = Source('http://mock.git', 'mock_dir')
+        source = Source('http://example.com/foo/bar.git')
 
-        assert 'http://mock.git' == source.repo
-        assert 'mock_dir' == source.name
+        assert 'http://example.com/foo/bar.git' == source.repo
+        assert 'bar' == source.name
         assert 'master' == source.rev
         assert None is source.link
 
     def test_init_rev(self):
         """Verify the revision can be customized."""
-        source = Source('http://mock.git', 'mock_dir', 'v1.0')
+        source = Source('http://mock.git', 'mock_name', 'v1.0')
 
         assert 'v1.0' == source.rev
 
     def test_init_link(self):
         """Verify the link can be set."""
-        source = Source('http://mock.git', 'mock_dir', link='mock/link')
+        source = Source('http://mock.git', 'mock_name', link='mock/link')
 
         assert 'mock/link' == source.link
 
     def test_init_error(self):
-        """Verify the repository and directory are required."""
+        """Verify the repository, name, and rev are required."""
         with pytest.raises(ValueError):
-            Source('', 'mock_dir')
+            Source('', name='mock_name', rev='master')
         with pytest.raises(ValueError):
-            Source('http://mock.git', '')
+            Source('http://mock.git', name='', rev='master')
+        with pytest.raises(ValueError):
+            Source('http://mock.git', name='mock_name', rev='')
 
     def test_repr(self, source):
         """Verify sources can be represented."""
@@ -62,8 +64,8 @@ class TestSource:
 
     def test_lt(self):
         sources = [
-            Source('zzz', '123'),
-            Source('bbb', '456'),
+            Source('http://github.com/owner/123.git'),
+            Source('bbb', name='456'),
             Source('ccc', '456'),
             Source('BBB', 'AAA'),
             Source('AAA', 'AAA'),